+++ Packages/com.unity.cloud.gltfast/Runtime/Scripts/GltfAsset.cs
@@ -126,7 +126,7 @@ namespace GLTFast
             ICodeLogger logger = null
             )
         {
-            logger = logger ?? new ConsoleLogger();
+            logger ??= ConsoleLogger.Instance;
             var success = await base.Load(gltfUrl, downloadProvider, deferAgent, materialGenerator, logger);
             if (success)
             {